requirements (1).txt
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.3.3
numpy>=1.19.4
Pillow>=8.0.1
scikit-learn>=0.23.2
pandas>=1.1.4
tqdm>=4.51.0
torchsummary>=1.5.1
